one more check added

+++ user-service/src/main/java/com/stanzaliving/user/acl/service/impl/UserDepartmentLevelRoleServiceImpl.java
@@ -53,7 +53,8 @@ public class UserDepartmentLevelRoleServiceImpl implements UserDepartmentLevelRo
 		userDepartmentLevelRoleDbService.delete(userDepartmentLevelRoleEntityListExisting);
 		if(CollectionUtils.isNotEmpty(userDepartmentLevelRoleEntityListExisting) && Objects.nonNull(userDepartmentLevelRoleEntityListExisting.get(0).getUserDepartmentLevelUuid())){
 			UserDepartmentLevelEntity departmentLevelEntity = userDepartmentLevelDbService.findByUuid(userDepartmentLevelRoleEntityListExisting.get(0).getUserDepartmentLevelUuid());
-			userDepartmentLevelDbService.delete(departmentLevelEntity);
+			if(Objects.nonNull(departmentLevelEntity))
+				userDepartmentLevelDbService.delete(departmentLevelEntity);
 		}
 	}
